Move split keyword func to `autofill` package
