+++ internal/autofill/split.go
@@ -1,4 +1,4 @@
-package parser
+package autofill
 
 import "strings"
