+++ internal/autofill/split_test.go
@@ -1,4 +1,4 @@
-package parser
+package autofill
 
 import (
 	"testing"
